pyproject.toml
[project]
dependencies = ["pyserial>=3.5", "cobs>=1.2.0", "rerun-sdk==0.30.2"]
name = "viz"
version = "0.1.0"

[project.gui-scripts]
# Alternative if you want to run it without a console (though this is a CLI)
# viz-gui = "viz.__main__:main"

[project.scripts]
viz = "viz.__main__:main"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]
